fix: only dispatch statechange on radio button panel
